Working version of the refactored observer
