+++ src/robot_control/nodes/observer
@@ -1,7 +1,52 @@
 #!/usr/bin/env python
 
-from robot_control.observer import IdealObs
+import rospy
+from robot_control.observer_lib import *
+from rospy import Subscriber, Publisher
+
+from geometry_msgs.msg import Pose2D
+from gazebo_msgs.msg import ModelStates
+
+class IdealObsNode:
+  ''' Ideal observer node
+  '''
+  def __init__(self):
+    rospy.init_node("observer")
+    rospy.sleep(0.5)
+
+    self.observer = IdealObs()
+    self.sensor_readings = {}
+
+    # Sensor topics
+    self.pose_sub = Subscriber("/gazebo/model_states",
+                               ModelStates,
+                               self.on_model_states)
+
+    self.state_pub = Publisher("state",
+                               Pose2D,
+                               queue_size=10)
+
+    self.run()
+
+  def on_model_states(self, state):
+    ''' Gazebo model state callback
+    '''
+    self.sensor_readings["true_state"] = state
+
+  def run(self):
+    ''' Main loop
+    '''
+    rate = rospy.Rate(10)
+    t_start = rospy.Time.now().to_sec()
+
+    while not rospy.is_shutdown():
+      state = self.observer.estimate(self.sensor_readings)
+      if state:
+        self.state_pub.publish(state)
+
+      rate.sleep()
+
 
 if __name__=="__main__":
-  IdealObs()
+  IdealObsNode()
